Create local ga() function
When the site is running locally, it won't load any of the js for google
analytics. The problem is that in my all.js script, in one of my
functions I'm sending data to google analytics. When running locally, I
get an error and it prevents the app from working properly.

That's why I added an emtpy ga() function that runs only locally.